Removing htmlproofer's HTML validation
htmlproofer's HTML validation doesn't seem to be playing nice with the AMP template we're using.
Removing it until it learns to behave itself or I find out what's happening.

PS.: As it is, the only thing being checked now is for the existence of favicon. So much for a CI, huh?